Moved utility classes to a common namespace Utilities. Split Vector2 and Rect2 into separate files. Cleaned up ObservableBase.
