+++ AnnoMapEditor/DataArchives/RdaDataArchive.cs
@@ -4,8 +4,8 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
+using AnnoMapEditor.Utilities;
 using Microsoft.Extensions.FileSystemGlobbing;
-using Microsoft.Extensions.FileSystemGlobbing.Abstractions;
 using RDAExplorer;
 
 namespace AnnoMapEditor.DataArchives
