+++ AnnoMapEditor/MapTemplates/IslandType.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using AnnoMapEditor.Utilities;
 using System.Collections.Generic;
 using System.Linq;
