+++ AnnoMapEditor/Mods/MapType.cs
@@ -1,4 +1,5 @@
 ﻿using AnnoMapEditor.MapTemplates;
+using AnnoMapEditor.Utilities;
 using System.Collections.Generic;
 using System.Linq;
