+++ AnnoMapEditor/UI/MainWindow.xaml.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Diagnostics;
 using System.Windows;
 using System.Windows.Controls;
+using AnnoMapEditor.Utilities;
 
 namespace AnnoMapEditor.UI
 {
