+++ AnnoMapEditor/UI/Models/ExportAsModViewModel.cs
@@ -1,8 +1,8 @@
-﻿using System.Collections.Generic;
+﻿using AnnoMapEditor.Utilities;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Windows;
 
 namespace AnnoMapEditor.UI.Models
 {
