+++ AnnoMapEditor/Utilities/ObservableBase.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel;
+using System.Runtime.CompilerServices;
+
+namespace AnnoMapEditor.Utilities
+{
+    public class ObservableBase : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler? PropertyChanged = delegate { };
+        protected void OnPropertyChanged([CallerMemberName] string? propertyName = null) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    }
+}
