+++ AnnoMapEditor/Utilities/Vector2.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 
-namespace AnnoMapEditor.MapTemplates
+namespace AnnoMapEditor.Utilities
 {
     public record Vector2
     {
@@ -104,59 +100,5 @@ namespace AnnoMapEditor.MapTemplates
         }
 
         private static int Normalize(int x) => (x + 4) / 8 * 8;
-    }
-
-    public struct Rect2
-    {
-        public Vector2 Position;
-        public Vector2 Size;
-        public Vector2 Max => Position + Size;
-
-        public int X => Position.X;
-        public int Y => Position.Y;
-        public int Width => Size.X;
-        public int Height => Size.Y;
-
-        public Rect2(string? area)
-        {
-            if (area is not null)
-            {
-                string[] parts = area.Split(' ');
-                if (parts.Length == 4)
-                {
-                    Position = new Vector2(int.Parse(parts[0]), int.Parse(parts[1]));
-                    Size = new Vector2(int.Parse(parts[2]), int.Parse(parts[3])) - Position;
-                    return;
-                }
-            }
-
-            Position = Vector2.Zero;
-            Size = Vector2.Zero;
-        }
-
-        public Rect2(int[]? numbers)
-        {
-            if (numbers?.Length == 4)
-            {
-                Position = new Vector2(numbers);
-                Size = new Vector2(numbers[2], numbers[3]) - Position;
-                return;
-            }
-
-            Position = Vector2.Zero;
-            Size = Vector2.Zero;
-        }
-
-        public Rect2(Vector2 max)
-        {
-            Position = Vector2.Zero;
-            Size = max - Position;
-        }
-
-        public Rect2(Vector2 min, Vector2 max)
-        {
-            Position = min;
-            Size = max - Position;
-        }
     }
 }
